Change parsing to fetch different substring from same string
DTB format might reuse a single string to represent different substrings.
In this case a substring must be the a suffix of an existing string.

For example "max_number" and "number" can be fetched from the
strings block from the same string that has "max_number". If
"max_number" is addressed by offset 27, "number" can be addressed by
the offset 31 using the same binary.